Removing basic amount of files to delete reindexer nonSingleton
