+++ orc8r/cloud/configs/orc8r_worker.yml
@@ -13,7 +13,3 @@
 # When true, state service handles automatically reindex state indexers.
 # When false, reindexing must be handled by the provided CLI.
 enable_automatic_reindexing: True
-
-# When true, use a singleton service to reindex states.
-# When false, use service(s) and a JobQueue to reindex states.
-enable_singleton_reindex: True
