+++ orc8r/cloud/configs/state.yml
@@ -9,13 +9,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-# TODO(reginawang3495): remove enable_automatic_reindexing and enable_singleton_reindex when jobQueue reindexing is removed
-
-# When true, state service handles automatically reindex state indexers.
-# When false, reindexing must be handled by the provided CLI.
-enable_automatic_reindexing: True
-
-# When true, use a singleton service to reindex states.
-# When false, use service(s) and a JobQueue to reindex states.
-enable_singleton_reindex: True
